Moving query_parser.py out of experiments. Looks like it's moving into the big leagues!
