+++ query_parser.py
@@ -3,7 +3,6 @@ from abc import ABC, abstractmethod
 from typing import List, Optional
 
 import pyparsing
-import pytest
 from pyparsing import Word, QuotedString, printables, Literal, Forward, ZeroOrMore, Group, \
     ParseResults, ParseException, CaselessLiteral, ParserElement
 from pyparsing.diagram import to_railroad, railroad_to_html
@@ -294,18 +293,18 @@ def query_parser() -> ParserElement:
     valid_chars = printables.replace("(", "").replace(")", "").replace(":", "").replace("\"", "")
     expr = Forward().setName("expression")
     quotes = QuotedString('"', "\\").setName("quoted string").setResultsName("quotes")
-    brackets = Group(Literal("(").suppress() + expr + Literal(")").suppress())\
+    brackets = Group(Literal("(").suppress() + expr + Literal(")").suppress()) \
         .setName("bracketed expression").setResultsName("brackets")
     words = Word(valid_chars).setName("word").setResultsName("word")
-    field_name = Group((Literal("@").suppress() + Word(valid_chars)) | (Word(valid_chars) + Literal(":").suppress()))\
+    field_name = Group((Literal("@").suppress() + Word(valid_chars)) | (Word(valid_chars) + Literal(":").suppress())) \
         .setName("field name").setResultsName("field_name")
     field_value = Group(quotes | words).setName("field value").setResultsName("field_value")
     field = Group(field_name + field_value).setName("field").setResultsName("field")
-    negator = Group(pyparsing.Optional(Literal("!") | Literal("-") | CaselessLiteral("not")))\
+    negator = Group(pyparsing.Optional(Literal("!") | Literal("-") | CaselessLiteral("not"))) \
         .setName("negator").setResultsName("negator")
     element = Group(quotes | brackets | field | words).setName("element").setResultsName("element")
     full_element = Group(negator + element).setName("full element").setResultsName("full_element", listAllMatches=True)
-    connector = Group(pyparsing.Optional(CaselessLiteral("or") | CaselessLiteral("and")))\
+    connector = Group(pyparsing.Optional(CaselessLiteral("or") | CaselessLiteral("and"))) \
         .setName("connector").setResultsName("connector", listAllMatches=True)
     expr <<= full_element + ZeroOrMore(connector + full_element)
     return expr
@@ -314,7 +313,7 @@ def query_parser() -> ParserElement:
 def create_railroad_diagram() -> None:
     expr = query_parser()
     # Creating railroad diagrams
-    with open("output.html", "w") as fp:
+    with open("experiments/output.html", "w") as fp:
         railroad = to_railroad(expr)
         fp.write(railroad_to_html(railroad))
 
@@ -417,54 +416,3 @@ def parse_word(word: str, field: Optional['Field'] = None) -> 'Query':
         regex = ".*".join(parts)
         return RegexQuery(regex, field)
     return WordQuery(word, field)
-
-
-def test_parser():
-    assert parse_query("first") == WordQuery("first")
-    assert parse_query("first document") == AndQuery([WordQuery("first"), WordQuery("document")])
-    assert parse_query("-first") == NotQuery(WordQuery("first"))
-    assert parse_query("!first") == NotQuery(WordQuery("first"))
-    assert parse_query("! first") == NotQuery(WordQuery("first"))
-    assert parse_query("not first") == NotQuery(WordQuery("first"))
-    assert parse_query("NOT first") == NotQuery(WordQuery("first"))
-    assert parse_query("first and document") == AndQuery([WordQuery("first"), WordQuery("document")])
-    assert parse_query("first or document") == OrQuery([WordQuery("first"), WordQuery("document")])
-    assert parse_query("first AND doc OR document") == OrQuery(
-        [AndQuery([WordQuery("first"), WordQuery("doc")]), WordQuery("document")]
-    )
-    assert parse_query("(first)") == WordQuery("first")
-    assert parse_query("first and (doc or document)") == AndQuery(
-        [WordQuery("first"), OrQuery([WordQuery("doc"), WordQuery("document")])]
-    )
-    assert parse_query("first (doc or document)") == AndQuery(
-        [WordQuery("first"), OrQuery([WordQuery("doc"), WordQuery("document")])]
-    )
-    with pytest.raises(InvalidQueryException):
-        parse_query("first (doc or document")
-    assert parse_query("\"first document\"") == PhraseQuery("first document")
-    assert parse_query("keyword:first document") == AndQuery(
-        [WordQuery("first", KeywordField()), WordQuery("document")]
-    )
-    assert parse_query("keyword:\"first document\"") == PhraseQuery("first document", KeywordField())
-    with pytest.raises(InvalidQueryException):
-        parse_query("keyword:(first and document)")
-    assert parse_query("@keyword first document") == AndQuery(
-        [WordQuery("first", KeywordField()), WordQuery("document")]
-    )
-    assert parse_query("title:first") == WordQuery("first", TitleField())
-    assert parse_query("TITLE: first") == WordQuery("first", TitleField())
-    assert parse_query("description:first") == WordQuery("first", DescriptionField())
-    assert parse_query("rating:general") == RatingQuery(Rating.GENERAL)
-    with pytest.raises(InvalidQueryException):
-        parse_query("fake:first")
-    assert parse_query("first*") == PrefixQuery("first")
-    assert parse_query("*first") == SuffixQuery("first")
-    assert parse_query("fi*st") == RegexQuery("fi.*st")
-    assert parse_query("fi[*st") == RegexQuery(r"fi\[.*st")
-    assert parse_query("\"Hello WORLD!\"") == PhraseQuery("Hello WORLD!")
-    assert parse_query("not (hello \"first :) document\")") == NotQuery(
-        AndQuery([WordQuery("hello"), PhraseQuery("first :) document")])
-    )
-    with pytest.raises(InvalidQueryException):
-        parse_query("\"hello \" document\"")
-    assert parse_query("\"hello \\\" document\"") == PhraseQuery("hello \" document")
